Fix the usage pages for simulation usages.
Getting the usage pages right is important, because DirectInput will make determinations about "device type" by looking at the usage-page/usage combination (which is not the only way to influence what "device type" DirectInput thinks we are, but is the easiest way for the time being). Some applications care about the device type.